+++ etna/models/nn/deepar.py
@@ -11,6 +11,7 @@ from etna.datasets.tsdataset import TSDataset
 from etna.loggers import tslogger
 from etna.models.base import Model
 from etna.models.base import log_decorator
+from etna.transforms import PytorchForecastingTransform
 
 
 class DeepARModel(Model):
@@ -89,6 +90,16 @@ class DeepARModel(Model):
             dropout=self.dropout,
         )
 
+    @staticmethod
+    def _get_pf_transform(ts: TSDataset) -> PytorchForecastingTransform:
+        """Get PytorchForecastingTransform from ts.transforms or raise exception if not found."""
+        if ts.transforms is not None and isinstance(ts.transforms[-1], PytorchForecastingTransform):
+            return ts.transforms[-1]
+        else:
+            raise ValueError(
+                "Not valid usage of transforms, please add PytorchForecastingTransform at the end of transforms"
+            )
+
     @log_decorator
     def fit(self, ts: TSDataset) -> "DeepARModel":
         """
@@ -103,7 +114,8 @@ class DeepARModel(Model):
         -------
         DeepARModel
         """
-        self.model = self._from_dataset(ts.transforms[-1].pf_dataset_train)
+        pf_transform = self._get_pf_transform(ts)
+        self.model = self._from_dataset(pf_transform.pf_dataset_train)
 
         self.trainer = pl.Trainer(
             logger=tslogger.pl_loggers,
@@ -113,7 +125,7 @@ class DeepARModel(Model):
             gradient_clip_val=self.gradient_clip_val,
         )
 
-        train_dataloader = ts.transforms[-1].pf_dataset_train.to_dataloader(train=True, batch_size=self.batch_size)
+        train_dataloader = pf_transform.pf_dataset_train.to_dataloader(train=True, batch_size=self.batch_size)
 
         self.trainer.fit(self.model, train_dataloader)
 
@@ -134,7 +146,8 @@ class DeepARModel(Model):
         TSDataset
             TSDataset with predictions.
         """
-        prediction_dataloader = ts.transforms[-1].pf_dataset_predict.to_dataloader(
+        pf_transform = self._get_pf_transform(ts)
+        prediction_dataloader = pf_transform.pf_dataset_predict.to_dataloader(
             train=False, batch_size=self.batch_size * 2
         )