+++ etna/transforms/__init__.py
@@ -11,6 +11,7 @@ from etna.transforms.lags import LagTransform
 from etna.transforms.log import LogTransform
 from etna.transforms.outliers import DensityOutliersTransform
 from etna.transforms.outliers import MedianOutliersTransform
+from etna.transforms.outliers import SAXOutliersTransform
 from etna.transforms.power import BoxCoxTransform
 from etna.transforms.power import YeoJohnsonTransform
 from etna.transforms.pytorch_forecasting import PytorchForecastingTransform