+++ tests/test_models/nn/test_deepar.py
@@ -8,8 +8,27 @@ from pytorch_forecasting.data import GroupNormalizer
 from etna.datasets.tsdataset import TSDataset
 from etna.metrics import MAE
 from etna.models.nn import DeepARModel
+from etna.transforms import AddConstTransform
 from etna.transforms import DateFlagsTransform
 from etna.transforms import PytorchForecastingTransform
+
+
+def test_fit_wrong_order_transform(weekly_period_df):
+    ts = TSDataset(TSDataset.to_dataset(weekly_period_df), "D")
+    add_const = AddConstTransform(in_column="target", value=1.0)
+    pft = PytorchForecastingTransform(
+        max_encoder_length=21,
+        max_prediction_length=8,
+        time_varying_known_reals=["time_idx"],
+        time_varying_unknown_reals=["target"],
+        target_normalizer=GroupNormalizer(groups=["segment"]),
+    )
+
+    ts.fit_transform([pft, add_const])
+
+    model = DeepARModel(max_epochs=300, learning_rate=[0.1])
+    with pytest.raises(ValueError, match="add PytorchForecastingTransform"):
+        model.fit(ts)
 
 
 @pytest.mark.long
@@ -31,8 +50,8 @@ def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
         weekly_period_df[lambda x: x.timestamp >= ts_start],
     )
 
-    ts_train = TSDataset(TSDataset.to_dataset(train), "1d")
-    ts_test = TSDataset(TSDataset.to_dataset(test), "1d")
+    ts_train = TSDataset(TSDataset.to_dataset(train), "D")
+    ts_test = TSDataset(TSDataset.to_dataset(test), "D")
     dft = DateFlagsTransform(day_number_in_week=True, day_number_in_month=False)
     pft = PytorchForecastingTransform(
         max_encoder_length=21,
@@ -45,10 +64,10 @@ def test_deepar_model_run_weekly_overfit(weekly_period_df, horizon):
 
     ts_train.fit_transform([dft, pft])
 
-    tftmodel = DeepARModel(max_epochs=300, learning_rate=[0.1])
+    model = DeepARModel(max_epochs=300, learning_rate=[0.1])
     ts_pred = ts_train.make_future(horizon)
-    tftmodel.fit(ts_train)
-    ts_pred = tftmodel.forecast(ts_pred)
+    model.fit(ts_train)
+    ts_pred = model.forecast(ts_pred)
 
     mae = MAE("macro")